requirements.txt
## AI Finance Assistant - Dependencies
## Core LLM & Agentic Framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

## Vector Database & RAG
faiss-cpu>=1.8.0
chromadb>=0.5.0

## Web Interface
streamlit>=1.38.0
plotly>=5.24.0

## Market Data
yfinance>=0.2.44
requests>=2.32.0

## Data Processing
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

## Observability
langfuse>=2.0.0

## Guardrails (optional)
# guardrails-ai>=0.5.0

## Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

## Utilities
pydantic>=2.8.0
httpx>=0.27.0
tenacity>=9.0.0
cachetools>=5.5.0
tavily-python>=0.3.0
